+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v2.1.0
 
+- Bugfix: computation of `trend` is changed to use balanced distance from diagonal [(#167)](https://github.com/JuliaDynamics/RecurrenceAnalysis.jl/issues/167)
+
+
 Updated to StateSpaceSets.jl v2.
 
 ## v2.0.0